Use panic() instead of log.Fatal() for better testability
Replace calls to log.Fatal by calling panic() with an error object
from cmd/errors.go. This makes it much easier to test things because
we can decide to exit with an error when running the CLI or capture
an error during testing based on how we implement a recover() method.