Add GitHub workflow to run pre-commit checks
This removes the prow-based CI machinery and replaces it with a
GitHub workflow (`.github/workflows/precommit.yml`) that runs on
pull requests and pushes.

The workflow runs a variety of linters as well as Go unit tests.

We take advantage of GitHub's caching support to improve the runtime
of the checks when the project dependencies haven't changed.

Closes #3

Change-Id: I38ee4ca830093199807140d3005214231e2b1dfa